Allow the default transition to be a float

+++ custom_components/dmx/light.py
@@ -140,7 +140,7 @@ PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend({
             vol.Optional(CONF_DEFAULT_COLOR): vol.All(
                 vol.ExactSequence((cv.byte, cv.byte, cv.byte)),
                 vol.Coerce(tuple)),
-            vol.Optional(CONF_TRANSITION, default=0): vol.All(vol.Coerce(int),
+            vol.Optional(CONF_TRANSITION, default=0): vol.All(vol.Coerce(float),
                                                               vol.Range(min=0,
                                                               max=60)),
             vol.Optional(CONF_CHANNEL_SETUP): cv.string,
